# api/openapi.yaml
openapi: 3.1.0
info:
  title: Rackd API
  description: Infrastructure management API for devices, networks, datacenters, and discovery
  version: 1.0.0
  contact:
    name: Rackd
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
security:
  - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
        default: 100
        maximum: 1000
    offsetParam:
      name: offset
      in: query
      schema:
        type: integer
        default: 0
    idPath:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  schemas:
    Error:
      type: object
      required: [code, error]
      properties:
        code:
          type: string
          description: >
            One of: INVALID_JSON, INVALID_INPUT, VALIDATION_ERROR, NOT_FOUND,
            FORBIDDEN, UNAUTHORIZED, INVALID_CREDENTIALS, ALREADY_EXISTS,
            IP_NOT_AVAILABLE, CANNOT_DELETE_SELF, SYSTEM_ROLE,
            RATE_LIMIT_EXCEEDED, LOGIN_RATE_LIMIT_EXCEEDED, CSRF_FAILED,
            EXPIRED_KEY, LEGACY_API_KEY_UNSUPPORTED, INTERNAL_ERROR
        error:
          type: string
        details:
          type: object
          additionalProperties: true
...
    User:
      type: object
      required: [id, username, is_active, is_admin, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        username: { type: string }
        email: { type: string }
        full_name: { type: string }
        is_active: { type: boolean }
        is_admin: { type: boolean }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        last_login_at: { type: string, format: date-time }
        roles:
          type: array
          items:
            $ref: '#/components/schemas/Role'
      additionalProperties: true
    UserInput:
      type: object
      required: [username, password]
      properties:
        username: { type: string }
        password: { type: string }
        email: { type: string }
        full_name: { type: string }
        is_admin: { type: boolean, default: false }
        role_id: { type: string, format: uuid }
    UserUpdate:
      type: object
      properties:
        username: { type: string }
        email: { type: string }
        full_name: { type: string }
        is_active: { type: boolean }
        is_admin: { type: boolean }
...
    Role:
      type: object
      required: [id, name, is_system, created_at, updated_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        description: { type: string }
        is_system: { type: boolean }
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
      additionalProperties: true
    RoleInput:
      type: object
      required: [name]
      properties:
        name: { type: string }
        description: { type: string }
        permissions:
          type: array
          items: { type: string }
    Permission:
      type: object
      required: [id, name, resource, action, created_at]
      properties:
        id: { type: string }
        name: { type: string }
        resource: { type: string }
        action: { type: string }
        created_at: { type: string, format: date-time }
...
  responses:
    BadRequest:
      description: Invalid input
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
paths:
...
  /api/users:
    get:
      operationId: listUsers
      tags: [Users]
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      operationId: createUser
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/users/{id}:
    parameters:
      - $ref: '#/components/parameters/idPath'
    get:
      operationId: getUser
      tags: [Users]
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      operationId: updateUser
      tags: [Users]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserUpdate'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      operationId: deleteUser
      tags: [Users]
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /api/roles:
    get:
      operationId: listRoles
      tags: [Roles]
      parameters:
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
      responses:
        '200':
          description: List of roles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Role'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      operationId: createRole
      tags: [Roles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleInput'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/roles/{id}:
    parameters:
      - $ref: '#/components/parameters/idPath'
    get:
      operationId: getRole
      tags: [Roles]
      responses:
        '200':
          description: Role details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    put:
      operationId: updateRole
      tags: [Roles]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleInput'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Role'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
    delete:
      operationId: deleteRole
      tags: [Roles]
      responses:
        '204': { description: Deleted }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
...
  /api/permissions:
    get:
      operationId: listPermissions
      tags: [Roles]
      responses:
        '200':
          description: All available permissions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Permission'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/InternalError' }

  # ── OAuth 2.1 ──
...